requirements.txt
Django>=4.0.1,<4.1.0
djangorestframework>=3.13.1,<3.14.0
django-environ>=0.8.1,<0.9.0
djangorestframework-simplejwt>=5.0.0,<5.1.0
django-cors-headers>=3.11.0,<3.12.0